backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic>=2.0
paho-mqtt>=2.1.0
